+++ app/lib/l10n/app_localizations_nl.dart
@@ -4110,6 +4110,17 @@ class AppLocalizationsNl extends AppLocalizations {
   }
 
   @override
+  String importReviewWillImportPrograms(int count) {
+    String _temp0 = intl.Intl.pluralLogic(
+      count,
+      locale: localeName,
+      other: '$count programma\'s',
+      one: '1 programma',
+    );
+    return 'Bevat ook $_temp0.';
+  }
+
+  @override
   String get importReviewCouldNotRead => 'De import kon niet worden gelezen';
 
   @override